+++ db/migrate/20161025190113_associate_creator_authors_with_their_user.rb
@@ -1,0 +1,4 @@
+class AssociateCreatorAuthorsWithTheirUser < DataMigration
+  RAKE_TASK_UP = 'data:migrate:authors:connect_creators_with_user_record'
+  RAKE_TASK_DOWN = 'data:migrate:authors:disconnect_creators_from_user_record'
+end